requirements.txt
django-jet
pyyaml 
ua-parser 
user-agents
django-user-agents
django-mptt
django-admin-sortable2